requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.1.1
streamlit==1.26.0
